refactor: extract menu sessions from App
