+++ src/app/App.tsx
@@ -324,6 +324,8 @@ import { useAppExternalEvents } from "./use-app-external-events";
 
 import { useAgentAttentionNotifications } from "./use-agent-attention-notifications";
 
+import { useAgentMenuSessions } from "./use-agent-menu-sessions";
+
 export function App() {
   const replayOnboarding = shouldReplayOnboarding();
   const activeHermesProfileName = useActiveHermesProfileName();
@@ -1860,71 +1862,10 @@ export function App() {
   // click-through notification per newly finished run. State persists so
   // reloads and restarts never renotify, and the first poll of an install
   // baselines silently instead of backfilling history.
-  useEffect(() => {
-    if (appBlocked || !bootstrapped) return;
-    let cancelled = false;
-    let state = loadRoutineRunWatchState();
-    const runSingleFlight = createSingleFlight();
-
-    async function notificationPermissionGranted() {
-      let granted = await isPermissionGranted().catch(() => false);
-      if (!granted) {
-        const permission = await requestPermission().catch(() => "denied" as const);
-        granted = permission === "granted";
-      }
-      return granted;
-    }
-
-    async function poll() {
-      await runSingleFlight(async () => {
-        let sessions: HermesSessionInfo[];
-        try {
-          sessions = await listScheduledRunSessions({ includeActive: true });
-        } catch {
-          // Bridge down (asleep, restarting): try again next tick.
-          return;
-        }
-        if (cancelled) return;
-        const { next, notices } = routineRunWatchStep(state, sessions, Date.now());
-        state = next;
-        if (notices.length === 0) {
-          saveRoutineRunWatchState(state);
-          return;
-        }
-        // Match agent/recording notification paths: ask for permission before
-        // sending, and only mark delivered after a successful send so a denied
-        // or failed delivery can retry while the run is still fresh.
-        if (!(await notificationPermissionGranted())) {
-          saveRoutineRunWatchState(state);
-          return;
-        }
-        if (cancelled) return;
-        const delivered: string[] = [];
-        await Promise.all(
-          notices.map((notice) =>
-            sendAppNotification({
-              title: notice.title,
-              body: notice.body,
-              sound: "Ping",
-              group: notice.jobId ? `june-routine-${notice.jobId}` : "june-routine",
-              sessionId: notice.sessionId,
-            })
-              .then(() => delivered.push(notice.sessionId))
-              .catch(() => {}),
-          ),
-        );
-        state = markRunsNotified(state, delivered);
-        saveRoutineRunWatchState(state);
-      });
-    }
-
-    void poll();
-    const timer = window.setInterval(() => void poll(), ROUTINE_RUN_NOTIFY_POLL_MS);
-    return () => {
-      cancelled = true;
-      window.clearInterval(timer);
-    };
-  }, [appBlocked, bootstrapped]);
+  useAgentMenuSessions({
+    appBlocked,
+    bootstrapped,
+  });
 
   // Project assignments for agent sessions, loaded once storage is up.
   useEffect(() => {
