+++ src/app/use-agent-menu-sessions-types.ts
@@ -1,0 +1,4 @@
+export type useAgentMenuSessionsDependencies = {
+  appBlocked: boolean;
+  bootstrapped: boolean;
+};
